+++ components/remote-controls/grid/remote/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/composable/RemoteGridComposable.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.remotecontrols.impl.grid.remote.composable
 
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.material.rememberScaffoldState
@@ -10,6 +11,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import com.flipperdevices.core.ui.theme.LocalPalletV2
+import com.flipperdevices.faphub.errors.api.FapHubComposableErrorsRenderer
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi
 import com.flipperdevices.remotecontrols.impl.grid.remote.composable.components.RemoteGridComposableContent
 import com.flipperdevices.remotecontrols.impl.grid.remote.composable.components.RemoteGridTopBar
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.decompose.RemoteGridComponent
@@ -17,6 +20,8 @@ import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.decompose
 @Composable
 fun RemoteGridComposable(
     remoteGridComponent: RemoteGridComponent,
+    flipperDispatchDialogApi: FlipperDispatchDialogApi,
+    errorsRenderer: FapHubComposableErrorsRenderer,
     modifier: Modifier = Modifier
 ) {
     val coroutineScope = rememberCoroutineScope()
@@ -42,8 +47,12 @@ fun RemoteGridComposable(
         content = { scaffoldPaddings ->
             RemoteGridComposableContent(
                 remoteGridComponent = remoteGridComponent,
+                flipperDispatchDialogApi = flipperDispatchDialogApi,
                 model = model,
-                modifier = Modifier.padding(scaffoldPaddings)
+                modifier = Modifier
+                    .padding(scaffoldPaddings)
+                    .navigationBarsPadding(),
+                errorsRenderer = errorsRenderer
             )
         }
     )