+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/composable/SetupScreen.kt
@@ -4,6 +4,7 @@ import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.togetherWith
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
@@ -14,10 +15,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import com.flipperdevices.core.ui.dialog.composable.busy.ComposableFlipperBusy
 import com.flipperdevices.core.ui.theme.LocalPalletV2
-import com.flipperdevices.ifrmvp.core.ui.layout.shared.ErrorComposable
+import com.flipperdevices.faphub.errors.api.FapErrorSize
+import com.flipperdevices.faphub.errors.api.FapHubComposableErrorsRenderer
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.SharedTopBar
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi
+import com.flipperdevices.remotecontrols.impl.setup.composable.components.AnimatedConfirmContent
 import com.flipperdevices.remotecontrols.impl.setup.composable.components.LoadedContent
 import com.flipperdevices.remotecontrols.impl.setup.composable.components.SetupLoadingContent
 import com.flipperdevices.remotecontrols.impl.setup.presentation.decompose.SetupComponent
@@ -29,7 +32,7 @@ import com.flipperdevices.remotecontrols.setup.impl.R as SetupR
 
 private val SetupComponent.Model.key: Any
     get() = when (this) {
-        SetupComponent.Model.Error -> "error"
+        is SetupComponent.Model.Error -> "error"
         is SetupComponent.Model.Loaded -> "loaded"
         is SetupComponent.Model.Loading -> "loading"
     }
@@ -38,6 +41,8 @@ private val SetupComponent.Model.key: Any
 @Composable
 fun SetupScreen(
     setupComponent: SetupComponent,
+    errorsRenderer: FapHubComposableErrorsRenderer,
+    flipperDispatchDialogApi: FlipperDispatchDialogApi,
     modifier: Modifier = Modifier
 ) {
     val rootNavigation = LocalRootNavigation.current
@@ -45,6 +50,7 @@ fun SetupScreen(
     val model by remember(setupComponent, coroutineScope) {
         setupComponent.model(coroutineScope)
     }.collectAsState()
+    val lastEmulatedSignal by setupComponent.lastEmulatedSignal.collectAsState()
     LaunchedEffect(setupComponent.remoteFoundFlow) {
         setupComponent.remoteFoundFlow.onEach {
             setupComponent.onFileFound(it)
@@ -73,25 +79,23 @@ fun SetupScreen(
             contentKey = { it.key }
         ) { model ->
             when (model) {
-                SetupComponent.Model.Error -> {
-                    ErrorComposable(onReload = setupComponent::onSuccessClick)
+                is SetupComponent.Model.Error -> {
+                    errorsRenderer.ComposableThrowableError(
+                        throwable = model.throwable,
+                        onRetry = setupComponent::tryLoad,
+                        fapErrorSize = FapErrorSize.FULLSCREEN,
+                        modifier = Modifier.fillMaxSize()
+                    )
                 }
 
                 is SetupComponent.Model.Loaded -> {
-                    if (model.isFlipperBusy) {
-                        ComposableFlipperBusy(
-                            onDismiss = setupComponent::dismissBusyDialog,
-                            goToRemote = {
-                                setupComponent.dismissBusyDialog()
-                                rootNavigation.push(RootScreenConfig.ScreenStreaming)
-                            }
-                        )
-                    }
+                    flipperDispatchDialogApi.Render(
+                        dialogType = model.flipperDialog,
+                        onDismiss = setupComponent::dismissDialog
+                    )
                     LoadedContent(
                         model = model,
                         modifier = Modifier.padding(scaffoldPaddings),
-                        onPositiveClick = setupComponent::onSuccessClick,
-                        onNegativeClick = setupComponent::onFailedClick,
                         onDispatchSignalClick = setupComponent::dispatchSignal
                     )
                 }
@@ -101,5 +105,15 @@ fun SetupScreen(
                 }
             }
         }
+        AnimatedConfirmContent(
+            lastEmulatedSignal = lastEmulatedSignal,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(scaffoldPaddings),
+            onNegativeClick = setupComponent::onFailedClick,
+            onSuccessClick = setupComponent::onSuccessClick,
+            onSkipClick = setupComponent::onSkipClicked,
+            onDismissConfirm = setupComponent::forgetLastEmulatedSignal
+        )
     }
 }