+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/presentation/decompose/SetupComponent.kt
@@ -1,10 +1,13 @@
 package com.flipperdevices.remotecontrols.impl.setup.presentation.decompose
 
 import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.faphub.errors.api.throwable.FapHubError
 import com.flipperdevices.ifrmvp.backend.model.IfrFileModel
+import com.flipperdevices.ifrmvp.backend.model.SignalResponse
 import com.flipperdevices.ifrmvp.backend.model.SignalResponseModel
 import com.flipperdevices.ifrmvp.model.IfrKeyIdentifier
 import com.flipperdevices.infrared.api.InfraredConnectionApi.InfraredEmulateState
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi
 import com.flipperdevices.remotecontrols.api.SetupScreenDecomposeComponent
 import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
 import kotlinx.coroutines.CoroutineScope
@@ -15,15 +18,18 @@ interface SetupComponent {
     fun model(coroutineScope: CoroutineScope): StateFlow<Model>
 
     val remoteFoundFlow: Flow<IfrFileModel>
+    val lastEmulatedSignal: StateFlow<SignalResponse?>
     val param: SetupScreenDecomposeComponent.Param
 
     fun onBackClick()
 
     fun onSuccessClick()
     fun onFailedClick()
+    fun onSkipClicked()
     fun dispatchSignal()
+    fun forgetLastEmulatedSignal()
 
-    fun dismissBusyDialog()
+    fun dismissDialog()
 
     fun tryLoad()
 
@@ -33,9 +39,8 @@ interface SetupComponent {
         data class Loading(val progress: Float) : Model
         data class Loaded(
             val response: SignalResponseModel,
-            val isFlipperBusy: Boolean = false,
+            val flipperDialog: FlipperDispatchDialogApi.DialogType? = null,
             val emulatedKeyIdentifier: IfrKeyIdentifier?,
-            val isEmulated: Boolean,
             val connectionState: InfraredEmulateState
         ) : Model {
             val isSyncing: Boolean = listOf(
@@ -45,7 +50,7 @@ interface SetupComponent {
             val isConnected = connectionState != InfraredEmulateState.NOT_CONNECTED
         }
 
-        data object Error : Model
+        data class Error(val throwable: FapHubError) : Model
     }
 
     interface Factory {