+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/presentation/viewmodel/DispatchSignalViewModel.kt
@@ -10,11 +10,14 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.ui.lifecycle.DecomposeViewModel
+import com.flipperdevices.faphub.target.api.FlipperTargetProviderApi
+import com.flipperdevices.faphub.target.model.FlipperTarget
 import com.flipperdevices.ifrmvp.model.IfrKeyIdentifier
 import com.flipperdevices.infrared.editor.core.model.InfraredRemote
 import com.flipperdevices.keyemulate.api.EmulateHelper
 import com.flipperdevices.keyemulate.exception.AlreadyOpenedAppException
 import com.flipperdevices.keyemulate.model.EmulateConfig
+import com.flipperdevices.keyemulate.tasks.CloseEmulateAppTaskHolder
 import com.flipperdevices.remotecontrols.api.DispatchSignalApi
 import com.flipperdevices.remotecontrols.impl.setup.encoding.ByteArrayEncoder
 import com.flipperdevices.remotecontrols.impl.setup.encoding.JvmEncoder
@@ -26,13 +29,16 @@ import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, DispatchSignalApi::class)
 class DispatchSignalViewModel @Inject constructor(
     private val emulateHelper: EmulateHelper,
-    private val serviceProvider: FlipperServiceProvider
+    private val serviceProvider: FlipperServiceProvider,
+    private val closeEmulateAppTaskHolder: CloseEmulateAppTaskHolder,
+    private val flipperTargetProviderApi: FlipperTargetProviderApi,
 ) : DecomposeViewModel(),
     FlipperBleServiceConsumer,
     LogTagProvider,
@@ -42,24 +48,34 @@ class DispatchSignalViewModel @Inject constructor(
     private val _state = MutableStateFlow<DispatchSignalApi.State>(DispatchSignalApi.State.Pending)
     override val state = _state.asStateFlow()
 
-    private val _isEmulated = MutableStateFlow(false)
-    override val isEmulated = _isEmulated.asStateFlow()
-
     private var latestDispatchJob: Job? = null
 
     override fun reset() {
         viewModelScope.launch {
             latestDispatchJob?.cancelAndJoin()
             _state.value = DispatchSignalApi.State.Pending
-            _isEmulated.value = false
         }
     }
 
     override fun dispatch(
         identifier: IfrKeyIdentifier,
         remotes: List<InfraredRemote>,
-        ffPath: FlipperFilePath
+        ffPath: FlipperFilePath,
+        onDispatched: () -> Unit
     ) {
+        when (flipperTargetProviderApi.getFlipperTarget().value) {
+            FlipperTarget.NotConnected -> {
+                _state.update { DispatchSignalApi.State.FlipperNotConnected }
+                return
+            }
+
+            FlipperTarget.Unsupported -> {
+                _state.update { DispatchSignalApi.State.FlipperNotSupported }
+                return
+            }
+
+            else -> Unit
+        }
         val i = remotes.indexOfFirst { remote ->
             when (identifier) {
                 is IfrKeyIdentifier.Name -> {
@@ -68,11 +84,6 @@ class DispatchSignalViewModel @Inject constructor(
 
                 is IfrKeyIdentifier.Sha256 -> {
                     val encoder = JvmEncoder(ByteArrayEncoder.Algorithm.SHA_256)
-                    identifier.hash == encoder.encode(remote.toByteArray())
-                }
-
-                is IfrKeyIdentifier.MD5 -> {
-                    val encoder = JvmEncoder(ByteArrayEncoder.Algorithm.MD5)
                     identifier.hash == encoder.encode(remote.toByteArray())
                 }
 
@@ -93,14 +104,18 @@ class DispatchSignalViewModel @Inject constructor(
             args = remote.name,
             index = i
         )
-        dispatch(config, identifier)
+        dispatch(config, identifier, onDispatched)
     }
 
     override fun dismissBusyDialog() {
         _state.value = DispatchSignalApi.State.Pending
     }
 
-    override fun dispatch(config: EmulateConfig, identifier: IfrKeyIdentifier) {
+    override fun dispatch(
+        config: EmulateConfig,
+        identifier: IfrKeyIdentifier,
+        onDispatched: () -> Unit
+    ) {
         if (latestDispatchJob?.isActive == true) return
         latestDispatchJob = viewModelScope.launch(Dispatchers.Main) {
             _state.emit(DispatchSignalApi.State.Pending)
@@ -119,7 +134,7 @@ class DispatchSignalViewModel @Inject constructor(
                             delay(DEFAULT_SIGNAL_DELAY)
                             emulateHelper.stopEmulate(this, serviceApi.requestApi)
                             _state.emit(DispatchSignalApi.State.Pending)
-                            _isEmulated.emit(true)
+                            onDispatched.invoke()
                         } catch (ignored: AlreadyOpenedAppException) {
                             _state.emit(DispatchSignalApi.State.FlipperIsBusy)
                         } catch (e: Exception) {
@@ -133,6 +148,13 @@ class DispatchSignalViewModel @Inject constructor(
     }
 
     override fun onServiceApiReady(serviceApi: FlipperServiceApi) = Unit
+
+    override fun onDestroy() {
+        if (_state.value is DispatchSignalApi.State.Emulating) {
+            closeEmulateAppTaskHolder.closeEmulateApp(serviceProvider, emulateHelper)
+        }
+        super<DecomposeViewModel>.onDestroy()
+    }
 
     companion object {
         private const val DEFAULT_SIGNAL_DELAY = 500L